이미지 삭제 NPE 해결

+++ secondboard/src/main/java/myweb/secondboard/controller/BoardApiController.java
@@ -5,10 +5,14 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.regex.Pattern;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import myweb.secondboard.domain.Board;
 import myweb.secondboard.domain.boards.BoardUploadFile;
+import myweb.secondboard.service.BoardLikeService;
+import myweb.secondboard.service.BoardReportService;
 import myweb.secondboard.service.BoardService;
 import myweb.secondboard.service.ImageService;
 import org.apache.commons.lang.StringUtils;
@@ -20,6 +24,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
 import org.thymeleaf.util.PatternUtils;
 
+@Slf4j
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/api/boards")
@@ -27,6 +32,7 @@ public class BoardApiController {
 
   private final BoardService boardService;
   private final ImageService imageService;
+
 
   @PostMapping("/boardDelete/{boardId}")
   public JSONObject boardDelete(@PathVariable("boardId") Long boardId) {
@@ -37,7 +43,6 @@ public class BoardApiController {
 
     System.out.println("content = " + content);
     //content =  <p><img src="/image/2" style="width: 25%;"><img src="/image/3" style="width: 25%;"><br></p>
-    List<String> ImageNumbers = new ArrayList<>();
 
     for (int i = 0; i < content.length(); i++) {
       String number = StringUtils.substringBetween(content, "/image/", "\"");
@@ -45,21 +50,19 @@ public class BoardApiController {
         break;
       }
 
-      String fileName = imageService.findById(Long.valueOf(number)).getFilePath();
-      System.out.println("fileName = " + fileName);
-      content = StringUtils.substringAfter(content, "/image/");
-      imageService.deleteById(Long.valueOf(number));
-      File file = new File(fileName);
-      file.delete();
-    }
+      if (imageService.findById(Long.valueOf(number)).isPresent()) {
+        String fileName = imageService.findById(Long.valueOf(number)).get().getFilePath();
+        System.out.println("fileName = " + fileName);
+        content = StringUtils.substringAfter(content, "/image/");
+        imageService.deleteById(Long.valueOf(number));
+        File file = new File(fileName);
+        file.delete();
+      }
+      else break;
 
-    for (String imageNumber : ImageNumbers) {
-      System.out.println("imageNumber = " + imageNumber);
     }
 
     boardService.deleteById(boardId);
-
-
     result.put("result", "success");
     return result;
   }
